requirements.txt
# Core UI Framework
PyQt5>=5.15.0
PyQt5-sip>=12.9.0

# Video Playback
python-vlc>=3.0.0

# Async HTTP Client
aiohttp>=3.8.0
aiofiles>=23.0.0

# Real-time Communication
signalrcore>=0.9.5

# HTTP Requests
requests>=2.28.0

# Image Processing
Pillow>=9.0.0

# Environment Variables
python-dotenv>=1.0.0

# JSON Schema Validation
jsonschema>=4.17.0

# Logging
python-json-logger>=2.0.0

# System Utilities (for WiFi management)
# Note: These are system packages, install via apt on Raspberry Pi:
# - network-manager (provides nmcli)
# - xserver-xorg
# - xinit
